test: move outbox metric tests to dedicated module
